LektieMail/SMS: Ignore sending to ID that isn't in DB
If a group contains a (new) id, listed in a [sms-<name>] in config file
but Lektier have not been fetched for the class with that ID.

Typically when preparing a new class, but the class have not been
included in the list of LektieIDs or not created in the skolekom system.

If fskintra.py -q --sms <name>: Only print warning if there should have
been sent to some mail/SMS recipients. If no -q: Give a warning even
when trying to send to an empty group.