Include reading of notes and attachments from lektie-pages
notes are finished, there is a problem with notes, see
pgLektieDB.insert_or_update_db.

attachments are not finished, only the path to the attachment is
included, but the attachemnt is not downloaded.

notes and attachements are included in LektieMail and LektieWeb.

General notes listed in ini-file are skipped, they are not shown in
LektieMail.
In LektieWeb a general note can be inserted in top of the page via
new (optional ini-file located in the webserver - see readme).

LektieSMS will not get notes or attachments, since they take up a lot of
chars and SMS chars are a bit expensive.